Move Sidebar component into a folder
This file was starting to get large and the addition of the opacity
slider would have made it larger.
