+++ src/app/src/components/Sidebar/index.js
@@ -1,0 +1,1 @@
+export { default } from './Sidebar';
