+++ src/app/src/components/Sidebar/style.js
@@ -1,0 +1,1 @@
+export const paddingLeft = 4;
